pyproject.toml
[tool.poetry]
name = "astrolog"
version = "0.1.0"
description = "High-level astrology methods based on Swiss Ephemeris"
authors = ["Ign. Aiom Ominous <aiom.ominous@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pyswisseph = "=2.10.02.0.dev1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
